# assessment-api-developer/docs/swagger.yaml
openapi: 3.0.0
info:
  title: Customers API
  version: 1.0.0
  description: API for managing customer data
paths:
  /api/customers:
    get:
      summary: Get all customers
      tags: ['Customers']
      operationId: GetAllCustomers
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Customer'
    post:
      summary: Add a new customer
      tags: ['Customers']
      operationId: AddCustomer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Customer'
      responses:
        '201':
          description: Customer created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'
        '400':
          description: Bad request
  /api/customers/{id}:
    get:
      summary: Get a customer by ID
      tags: ['Customers']
      operationId: GetCustomer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'
        '404':
          description: Customer not found
    put:
      summary: Update an existing customer
      tags: ['Customers']
      operationId: UpdateCustomer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Customer'
      responses:
        '204':
          description: NoContent on successfull execution
        '400':
          description: Bad request
        '404':
          description: Customer not found
    delete:
      summary: Delete a customer
      tags: ['Customers']
      operationId: DeleteCustomer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Customer deleted successfully
        '404':
          description: Customer not found
components:
  schemas:
    Customer:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        address:
          type: string
        email:
          type: string
        phone:
          type: string
        city:
          type: string
        state:
          type: string
        zip:
          type: string
        country:
          type: string
        notes:
          type: string
        contactName:
          type: string
        contactPhone:
          type: string
        contactEmail:
          type: string
        contactTitle:
          type: string
        contactNotes:
          type: string